Create script.js
When on IMDB click this small button to search for the movie/series on RARBG.

The button is small and super lightweight. It always stays out of your way on the top left corner. Just click on it and a new tab will open.


If it doesn't work the RARBG url has probably changed. Just let me know and I'll update it.